+++ testsuite/compat/src/test/java/org/jboss/as/test/compat/jpa/hibernate/transformer/VerifyHibernate51CompatibilityJDSEnabledTransformerTestCase.java
@@ -24,22 +24,20 @@ package org.jboss.as.test.compat.jpa.hibernate.transformer;
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.as.arquillian.api.ServerSetup;
 import org.jboss.as.test.shared.util.AssumeTestGroupUtil;
 import org.jboss.shrinkwrap.api.Archive;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.EnterpriseArchive;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.BeforeClass;
-import org.junit.Ignore;
 import org.junit.runner.RunWith;
 
 /**
  * Enable Hibernate bytecode transformer for application with jboss-deployment-structure.xml
  */
 @RunWith(Arquillian.class)
-@Ignore // WFLY-11991 added a StateType#internalNullSafeGet that is not supported by default
-        // Hibernate51TransformSessImplMtds must be set to true with that change, so we only allow
-        // the tests that allow system properties to be specified to run (all of the same tests otherwise run).
+@ServerSetup({AbstractVerifyHibernate51CompatibilityTestCase.EnableSessImplMtdsSetupTask.class})
 public class VerifyHibernate51CompatibilityJDSEnabledTransformerTestCase
         extends AbstractVerifyHibernate51CompatibilityTestCase {
 
@@ -60,7 +58,7 @@ public class VerifyHibernate51CompatibilityJDSEnabledTransformerTestCase
         ear.addAsModule(war);
 
         ear.addAsManifestResource(VerifyHibernate51CompatibilityJDSEnabledTransformerTestCase.class.getPackage(),
-                "jboss-deployment-structure.xml","jboss-deployment-structure.xml");
+                "jboss-deployment-structure.xml", "jboss-deployment-structure.xml");
         return ear;
     }
 }
